added blocking/nonblocking read for posix compatible systems
I've added the read_blocking function so that one can block until input
is available. The previous read implementation was blocking on windows
but nonblocking on posix, since windows was not supported until yesterday
this is not a breaking change.

Along with that, I've added hide_cursor(false) to .Restored mode so that
the terminal is properly restored.