removed step_cursor, improve coord resolution and window isolation precursor
I've removed the `step_cursor` function because it hard to properly
implement without integer overflow causing it to be buggy, so now
only `move_cursor`.

I've improved the coordinate resolution by adding modulo with term_size
this prevents the weird behavior that happens when the cursor coordinates
goes outside the visible range, this behavior was also replicated on
windows terminal and various terminals on linux.

Added a precursor to window isolation, we can now know where cursor is
going to be given a certain string. Next will be to take that information
and simulate it when using a `Window`