Cbreak and Raw modes now activate alternate buffer
On windows unless an alternate buffer is enable the main buffer
will get spammed and create a huge scroll history. So to adhere
to what everyone else is doing. When Cbreak and Raw modes are
activated we also use the alternate buffer and go back to main
on exit.

Mouse input is now also only activated by default on raw mode and
has to manually be enabled it on cbreak mode. This is because if the
user interrupts the process and the process does not gracefully exit
this will result in a spam of SGR escapes and on windows in specific
this causes the shell to crash.